product - move the interfaces into base layer for reuse in sandbox
